+++ android/src/main/java/com/jiguang/jpush/JPushPlugin.java
@@ -122,6 +122,8 @@ public class JPushPlugin implements FlutterPlugin, MethodCallHandler {
             setWakeEnable(call, result);
         } else if (call.method.equals("setAuth")) {
             setAuth(call, result);
+        } else if (call.method.equals("testCountryCode")) {
+            testCountryCode(call, result);
         } else {
             result.notImplemented();
         }
@@ -136,8 +138,12 @@ public class JPushPlugin implements FlutterPlugin, MethodCallHandler {
         if (enable == null) {
             enable = false;
         }
-        Log.d("debug_110","setauth="+enable);
         JCollectionAuth.setAuth(context,enable);
+    }
+    private void testCountryCode(MethodCall call, Result result){
+        String code = call.arguments();
+        Log.d(TAG,"testCountryCode code="+code);
+        JCoreInterface.testCountryCode(code);
     }
     private void setWakeEnable(MethodCall call, Result result) {
         HashMap<String, Object> map = call.arguments();
@@ -191,7 +197,9 @@ public class JPushPlugin implements FlutterPlugin, MethodCallHandler {
 
     public void scheduleCache() {
         Log.d(TAG, "scheduleCache:");
-
+        if(JPushPlugin.instance==null||JPushPlugin.instance.channel==null){
+            return;
+        }
         List<Object> tempList = new ArrayList<Object>();
 
         if (dartIsReady) {
@@ -478,7 +486,8 @@ public class JPushPlugin implements FlutterPlugin, MethodCallHandler {
     static void transmitMessageReceive(String message, Map<String, Object> extras) {
         Log.d(TAG, "transmitMessageReceive " + "message=" + message + "extras=" + extras);
 
-        if (instance == null) {
+        if (instance == null||instance.channel==null) {
+            Log.d("JPushPlugin", "the instance is null");
             return;
         }
         Map<String, Object> msg = new HashMap<>();
@@ -497,7 +506,7 @@ public class JPushPlugin implements FlutterPlugin, MethodCallHandler {
         notification.put("extras", extras);
         JPushPlugin.openNotificationCache.add(notification);
 
-        if (instance == null) {
+        if (instance == null||instance.channel==null) {
             Log.d("JPushPlugin", "the instance is null");
             return;
         }
@@ -512,8 +521,8 @@ public class JPushPlugin implements FlutterPlugin, MethodCallHandler {
 
     public static void onNotifyMessageUnShow( NotificationMessage notificationMessage) {
         Log.e(TAG,"[onNotifyMessageUnShow] message:"+notificationMessage);
-        if (instance == null) {
-            Log.d(TAG, "the instance is null");
+        if (instance == null||instance.channel==null) {
+            Log.d("JPushPlugin", "the instance is null");
             return;
         }
         Map<String, Object> notification= new HashMap<>();
@@ -560,7 +569,8 @@ public class JPushPlugin implements FlutterPlugin, MethodCallHandler {
     static void transmitNotificationReceive(String title, String alert, Map<String, Object> extras) {
         Log.d(TAG, "transmitNotificationReceive " + "title=" + title + "alert=" + alert + "extras=" + extras);
 
-        if (instance == null) {
+        if (instance == null||instance.channel==null) {
+            Log.d("JPushPlugin", "the instance is null");
             return;
         }